Remove RavenTiming to work with WCF serialization
StackExchange.Profiling.Wcf won't work with
StackExchange.Profiling.RavenDb due to serialization problems arising
from RavenTiming inheriting from CustomTiming. Without using the
KnownType attribute on CustomTiming WCF serialization of RavenTiming
will fail. Since the inheritance isn't strictly necessary, I've removed
it for the sake of WCF compatibility.
